+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/sourcecontext/BundleSourcesTask.kt
@@ -5,6 +5,7 @@ import io.sentry.android.gradle.autoinstall.SENTRY_GROUP
 import io.sentry.android.gradle.sourcecontext.GenerateBundleIdTask.Companion.SENTRY_BUNDLE_ID_PROPERTY
 import io.sentry.android.gradle.telemetry.SentryTelemetryService
 import io.sentry.android.gradle.util.PropertiesUtil
+import io.sentry.android.gradle.util.asSentryCliExec
 import io.sentry.android.gradle.util.info
 import io.sentry.gradle.common.SentryVariant
 import java.io.File
@@ -30,7 +31,16 @@ abstract class BundleSourcesTask : Exec() {
     init {
         group = SENTRY_GROUP
         description = "Creates a Sentry source bundle file."
+
+        @Suppress("LeakingThis")
+        onlyIf {
+            includeSourceContext.getOrElse(false) &&
+                !sourceDir.asFileTree.isEmpty
+        }
     }
+
+    @get:Input
+    abstract val includeSourceContext: Property<Boolean>
 
     @get:PathSensitive(PathSensitivity.RELATIVE)
     @get:InputDirectory
@@ -64,6 +74,10 @@ abstract class BundleSourcesTask : Exec() {
     @get:Input
     @get:Optional
     abstract val sentryAuthToken: Property<String>
+
+    @get:Input
+    @get:Optional
+    abstract val sentryUrl: Property<String>
 
     @get:Internal
     abstract val sentryTelemetryService: Property<SentryTelemetryService>
@@ -104,6 +118,11 @@ abstract class BundleSourcesTask : Exec() {
 
         if (debug.getOrElse(false)) {
             args.add("--log-level=debug")
+        }
+
+        sentryUrl.orNull?.let {
+            args.add("--url")
+            args.add(it)
         }
 
         args.add("debug-files")
@@ -147,6 +166,7 @@ abstract class BundleSourcesTask : Exec() {
             sentryOrg: Provider<String>,
             sentryProject: Provider<String>,
             sentryAuthToken: Property<String>,
+            sentryUrl: Property<String>,
             includeSourceContext: Property<Boolean>,
             sentryTelemetryProvider: Provider<SentryTelemetryService>,
             taskSuffix: String = ""
@@ -159,6 +179,7 @@ abstract class BundleSourcesTask : Exec() {
                 task.sentryOrganization.set(sentryOrg)
                 task.sentryProject.set(sentryProject)
                 task.sentryAuthToken.set(sentryAuthToken)
+                task.sentryUrl.set(sentryUrl)
                 task.sourceDir.set(collectSourcesTask.flatMap { it.output })
                 task.cliExecutable.set(cliExecutable)
                 SentryPropertiesFileProvider.getPropertiesFilePath(project, variant)?.let {
@@ -166,12 +187,10 @@ abstract class BundleSourcesTask : Exec() {
                 }
                 task.bundleIdFile.set(generateDebugIdTask.flatMap { it.outputFile })
                 task.output.set(output)
-                task.onlyIf {
-                    includeSourceContext.getOrElse(false) &&
-                        !task.sourceDir.asFileTree.isEmpty
-                }
+                task.includeSourceContext.set(includeSourceContext)
                 task.sentryTelemetryService.set(sentryTelemetryProvider)
                 task.usesService(sentryTelemetryProvider)
+                task.asSentryCliExec()
             }
         }
     }