+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/sourcecontext/SourceContext.kt
@@ -55,6 +55,7 @@ class SourceContext {
                 sentryOrg?.let { project.provider { it } } ?: extension.org,
                 sentryProject?.let { project.provider { it } } ?: extension.projectName,
                 extension.authToken,
+                extension.url,
                 extension.includeSourceContext,
                 sentryTelemetryProvider,
                 taskSuffix
@@ -70,6 +71,7 @@ class SourceContext {
                 sentryOrg?.let { project.provider { it } } ?: extension.org,
                 sentryProject?.let { project.provider { it } } ?: extension.projectName,
                 extension.authToken,
+                extension.url,
                 extension.includeSourceContext,
                 taskSuffix
             )