+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/integration/SentryPluginWithFirebaseTest.kt
@@ -46,7 +46,7 @@ class SentryPluginWithFirebaseTest :
             .appendArguments("app:assembleRelease")
             .build()
 
-        assertTrue { "BUILD SUCCESSFUL" in result.output }
+        assertTrue(result.output) { "BUILD SUCCESSFUL" in result.output }
     }
 
     override val additionalBuildClasspath: String =